Add isLoggedIn property to sidebar component and navigation
The isLoggedIn property was added to the sidebar component as an input. This property was then used to modify the HTML of the sidebar so that certain options are only visible if the user is logged in. As a result, changes have also been made on the app component, now holding the isLoggedIn state and passing it to the sidebar component. A subscription has been set up to the authentication selector to track the logged in status of the user.