Update toast and registration components; refactor register service
This commit mainly updates toast and registration inputs, and revamps the register service by refactoring the way toast messages are displayed. It also incorporates toast and register components into sidebar and main components, alongside updating user authentication actions and effects to accommodate the changes. Some fixes have been applied to the registration form for a correct data input and the register service now uses the toastService instead of the toastWrapperService.